fix(security): prevent XSS via case-insensitive </script> and </style> injection
The previous `escape_script_content` used a case-sensitive `replace("</script", ...)`
but HTML parsers match end tags case-insensitively. Payloads like `</SCRIPT>` or
`</Script>` bypassed the escape and could break out of inline <script> blocks
(e.g. __REX_DATA__, __REX_MANIFEST__) when user-controlled data was reflected
via getServerSideProps.

Fix: replace all `<` with `\u003c` in script/style contexts (same approach as
Next.js `htmlEscapeJsonString` and `serialize-javascript`). This eliminates the
entire class of tag-closing attacks regardless of case.

Also fixes:
- Missing `escape_style_content` call in `assemble_document` (non-streaming path
  inlined CSS without escaping, while the streaming path did escape).
- Extracts document tests to `document_tests.rs` to stay under 700-line limit.
